ed448_goldilocks::curve::twedwards::extended

Struct ExtendedPoint

pub struct ExtendedPoint {
    pub(crate) X: FieldElement,
    pub(crate) Y: FieldElement,
    pub(crate) Z: FieldElement,
    pub(crate) T: FieldElement,
}

Fields
X: FieldElementY: FieldElementZ: FieldElementT: FieldElement
Implementations

impl ExtendedPoint

pub const GENERATOR: ExtendedPoint

Generator for the prime subgroup

pub const IDENTITY: ExtendedPoint

Identity point

pub fn add_extended(&self, other: &ExtendedPoint) -> ExtensiblePoint

Adds an extensible point to an extended point Returns an extensible point (3.1) https://iacr.org/archive/asiacrypt2008/53500329/53500329.pdf

pub fn sub_extended(&self, other: &ExtendedPoint) -> ExtensiblePoint

Subtracts an extensible point from an extended point Returns an extensible point This is a direct modification of the addition formula to the negation of other

pub fn add_affine_niels(&self, other: AffineNielsPoint) -> ExtensiblePoint

Adds an extensible point to an AffineNiels point Returns an Extensible point

pub fn add_projective_niels( &self, other: &ProjectiveNielsPoint, ) -> ExtensiblePoint

Adds an extensible point to a ProjectiveNiels point Returns an extensible point (3.1)[Last set of formulas] https://iacr.org/archive/asiacrypt2008/53500329/53500329.pdf This differs from the formula above by a factor of 2. Saving 1 Double Cost 8M

pub fn to_extensible(self) -> ExtensiblePoint

Converts an ExtendedPoint to an ExtensiblePoint

pub fn to_untwisted(self) -> EdwardsExtendedPoint

Uses a 2-isogeny to map the point to the Ed448-Goldilocks

pub fn to_projective_niels(self) -> ProjectiveNielsPoint

Converts an Extensible point to a ProjectiveNiels Point

pub(crate) fn is_on_curve(&self) -> Choice

Checks if the point is on the curve

pub fn negate(&self) -> ExtendedPoint

Negates a point

pub fn torque(&self) -> ExtendedPoint

Torques a point