ed448_goldilocks::decaf::points

Struct DecafPoint

pub struct DecafPoint(pub(crate) ExtendedPoint);

A Decaf point in the Twisted Edwards curve

Tuple Fields
0: ExtendedPoint
Implementations

impl DecafPoint

pub const GENERATOR: DecafPoint

The generator point

pub const IDENTITY: DecafPoint

The identity point

pub fn is_identity(&self) -> Choice

Check if the point is the identity

pub fn add(&self, other: &DecafPoint) -> DecafPoint

Add two points

pub fn sub(&self, other: &DecafPoint) -> DecafPoint

Subtract two points

pub fn compress(&self) -> CompressedDecaf

Compress this point

pub fn from_uniform_bytes(bytes: &[u8; 112]) -> Self

Construct a DecafPoint from 112 bytes of data.

If the input bytes are uniformly distributed, the resulting point will be uniformly distributed over the group, and its discrete log with respect to other points is unknown.

Implements map to curve according see https://datatracker.ietf.org/doc/rfc9380/ section 5.3.4 by splitting the input into two 56-byte halves, then applies the decaf448_map to each, and adds the results.

pub fn random(rng: impl CryptoRng) -> Self

👎Deprecated since 0.14.0:
use the Generate trait instead

DEPRECATED: Return a DecafPoint chosen uniformly at random using a user-provided RNG.

Use the Generate trait instead.