ed448_goldilocks

Module edwards

Re-exports

pub use affine::AffinePoint;

pub use affine::CompressedEdwardsY;

pub use extended::EdwardsPoint;

pub use scalar::EdwardsScalar;

pub use scalar::EdwardsScalarBytes;

pub use scalar::WideEdwardsScalarBytes;

Modules

affine 🔒

This module contains the code for the Goldilocks curve. The goldilocks curve is the (untwisted) Edwards curve with affine equation x^2 + y^2 = 1 - 39081x^2y^2 Scalar Multiplication for this curve is pre-dominantly delegated to the Twisted Edwards variation using a (doubling) isogeny Passing the point back to the Goldilocks curve using the dual-isogeny clears the cofactor. The small remainder of the Scalar Multiplication is computed on the untwisted curve. See https://www.shiftleft.org/papers/isogeny/isogeny.pdf for details

extended 🔒

scalar 🔒