ed448_goldilocks::edwards::affine

Struct CompressedEdwardsY

pub struct CompressedEdwardsY(pub [u8; 57]);

Represents a point on the Compressed Twisted Edwards Curve in little endian format where the most significant bit is the sign bit and the remaining 448 bits represent the y-coordinate

Tuple Fields
0: [u8; 57]
Implementations

impl CompressedEdwardsY

pub const GENERATOR: Self

The compressed generator point

pub const IDENTITY: Self

The compressed identity point

pub fn decompress_unchecked(&self) -> CtOption<AffinePoint>

Attempt to decompress to an AffinePoint.

Returns None if the input is not the \(y\)-coordinate of a curve point.

pub fn decompress(&self) -> CtOption<AffinePoint>

Attempt to decompress to an AffinePoint.

Returns None:

if the input is not the \(y\)-coordinate of a curve point.

if the input point is not on the curve.

if the input point has nonzero torsion component.

pub const fn as_bytes(&self) -> &[u8; 57]

View this CompressedEdwardsY as an array of bytes.

pub const fn to_bytes(&self) -> [u8; 57]

Copy this CompressedEdwardsY to an array of bytes.