ed448_goldilocks::edwards::extended

Struct EdwardsPoint

pub struct EdwardsPoint {
    pub(crate) X: FieldElement,
    pub(crate) Y: FieldElement,
    pub(crate) Z: FieldElement,
    pub(crate) T: FieldElement,
}

Represent points on the (untwisted) edwards curve using Extended Homogenous Projective Co-ordinates (x, y) -> (X/Z, Y/Z, Z, T) a = 1, d = -39081 XXX: Make this more descriptive Should this be renamed to EdwardsPoint so that we are consistent with Dalek crypto? Necessary as ExtendedPoint is not regular lingo?

Fields
X: FieldElementY: FieldElementZ: FieldElementT: FieldElement
Implementations

impl EdwardsPoint

pub const GENERATOR: Self = GOLDILOCKS_BASE_POINT

Generator for the prime subgroup

pub const IDENTITY: Self

Identity point

pub fn to_montgomery(&self) -> MontgomeryPoint

Convert this point to MontgomeryPoint

pub fn scalar_mul(&self, scalar: &EdwardsScalar) -> Self

Generic scalar multiplication to compute s*P

pub fn add(&self, other: &EdwardsPoint) -> Self

Add two points

pub fn double(&self) -> Self

Double this point

pub fn is_on_curve(&self) -> Choice

Check if this point is on the curve

pub fn to_affine(&self) -> AffinePoint

Convert this point to an AffinePoint.

pub(crate) fn to_twisted(self) -> TwistedExtensiblePoint

pub fn negate(&self) -> Self

Compute the negation of this point’s x-coordinate.

pub fn torque(&self) -> Self

Compute the negation of this point’s y-coordinate.

pub fn is_torsion_free(&self) -> Choice

Determine if this point is “torsion-free”, i.e., is contained in the prime-order subgroup.

Return

true if self has zero torsion component and is in the prime-order subgroup;

false if self has a nonzero torsion component and is not in the prime-order subgroup.