ed448_goldilocks::edwards

Module scalar

Type Aliases

EdwardsScalar

Ed448 scalar field.

EdwardsScalarBytes

The number of bytes needed to represent the scalar field

WideEdwardsScalarBytes

The number of bytes needed to represent the safely create a scalar from a random bytes