ed448_goldilocks::field::element

Struct FieldElement

pub struct FieldElement(pub(crate) ConstMontyForm<MODULUS, { MODULUS::LIMBS }>);

Tuple Fields
0: ConstMontyForm<MODULUS, { MODULUS::LIMBS }>
Implementations

impl FieldElement

pub const A_PLUS_TWO_OVER_FOUR: Self

pub const DECAF_FACTOR: Self

pub const EDWARDS_D: Self

pub const J: Self

pub const MINUS_ONE: Self

pub const NEG_EDWARDS_D: Self

pub const NEG_FOUR_TIMES_TWISTED_D: Self

pub const ONE: Self

pub const TWO: Self

pub const TWISTED_D: Self

pub const TWO_TIMES_TWISTED_D: Self

pub const Z: Self

pub const ZERO: Self

pub fn is_negative(&self) -> Choice

pub fn is_zero(&self) -> Choice

pub fn invert(&self) -> Self

Inverts a field element

pub fn square(&self) -> Self

fn square_n<const N: u32>(&self) -> FieldElement

Squares a field element n times

pub fn is_square(&self) -> Choice

pub fn unchecked_sqrt(&self) -> FieldElement

pub fn to_bytes(self) -> [u8; 56]

pub fn to_bytes_extended(self) -> [u8; 57]

pub fn from_bytes(bytes: &[u8; 56]) -> Self

pub fn from_bytes_extended(bytes: &[u8; 57]) -> Self

pub fn from_repr(bytes: &[u8; 56]) -> CtOption<Self>

pub fn double(&self) -> Self

pub(crate) fn inverse_square_root(&self) -> (FieldElement, Choice)

Computes the inverse square root of a field element Returns the result and a boolean to indicate whether self was a Quadratic residue

pub(crate) fn sqrt_ratio( u: &FieldElement, v: &FieldElement, ) -> (FieldElement, Choice)

Computes the square root ratio of two elements

pub(crate) fn div_by_2(&self) -> FieldElement

pub(crate) fn map_to_curve_elligator2(&self) -> AffinePoint

pub(crate) fn map_to_curve_decaf448(&self) -> TwistedExtendedPoint