ed448_goldilocks::field

Module scalar

Structs

Scalar

Shared scalar for Ed448 and Decaf448. Use EdwardsScalar and DecafScalar directly.

Constants

HALF_ORDER 🔒

MODULUS_LIMBS

The modulus of the scalar field as a sequence of 14 32-bit limbs

ORDER

The order of the scalar field

ORDER_MINUS_ONE 🔒

The order of the scalar field

WIDE_ORDER

The wide order of the scalar field

WIDE_ORDER_MINUS_ONE 🔒

Traits

CurveWithScalar

Type Aliases

ScalarBytes

The number of bytes needed to represent the scalar field

WideScalarBytes

The number of bytes needed to represent the safely create a scalar from a random bytes