ed448_goldilocks::sign::signing_key

Struct SigningKey

pub struct SigningKey {
    pub(crate) secret: ExpandedSecretKey,
}

Signing key for Ed448

Fields
secret: ExpandedSecretKey
Implementations

impl SigningKey

pub fn to_bytes(&self) -> SecretKey

Serialize this SigningKey as bytes.

pub fn as_bytes(&self) -> &SecretKey

Serialize this SigningKey as a byte reference.

pub fn to_scalar(&self) -> EdwardsScalar

Return the clamped EdwardsScalar for this SigningKey.

This is the scalar that is actually used for signing. Be warned, this is secret material that should be handled with care.

pub fn verifying_key(&self) -> VerifyingKey

Get the VerifyingKey for this SigningKey.

pub fn with_context<'k, 'v>( &'k self, context: &'v [u8], ) -> Context<'k, 'v, Self>

Create a signing context that can be used for Ed448ph with signature::DigestSigner

pub fn sign_raw(&self, message: &[u8]) -> Signature

Sign a message with this SigningKey using the Ed448 algorithm defined in RFC8032 5.2.

pub fn sign_ctx( &self, context: &[u8], message: &[u8], ) -> Result<Signature, Error>

Sign a message in the given context with this SigningKey using the Ed448ph algorithm defined in RFC8032 5.2.

pub fn sign_prehashed<D>( &self, context: Option<&[u8]>, prehashed_message: D, ) -> Result<Signature, Error>
where D: PreHash,

Sign a prehashed_message with this SigningKey using the Ed448ph algorithm defined in RFC8032 5.2.