ed448_goldilocks

Struct VerifyingKey

pub struct VerifyingKey {
    pub(crate) compressed: CompressedEdwardsY,
    pub(crate) point: EdwardsPoint,
}

Ed448 public key as defined in [RFC8032  5.2.5]

Fields
compressed: CompressedEdwardsYpoint: EdwardsPoint
Implementations

impl VerifyingKey

pub fn to_bytes(&self) -> [u8; 57]

Convert this verifying key into byte slice

pub fn as_bytes(&self) -> &[u8; 57]

View this public key as a byte slice.

pub fn from_bytes(bytes: &[u8; 57]) -> Result<Self, Error>

Construct a VerifyingKey from a slice of bytes.

pub fn with_context<'k, 'v>( &'k self, context: &'v [u8], ) -> Context<'k, 'v, Self>

Create a context for this verifying key that can be used with signature::DigestVerifier.

pub fn to_edwards(self) -> EdwardsPoint

Return the verifying key in Edwards form.

pub fn verify_raw( &self, signature: &Signature, message: &[u8], ) -> Result<(), Error>

Verifies a signature on a message.

This is the “Ed448” mode of RFC 8032 (no pre-hashing, a context is provided). This is equivalent to verify_ctx() with an empty (zero-length) context.

Note: this function is not constant-time; it assumes that the public key and signature value are public data.

pub fn verify_ctx( self, sig: &Signature, ctx: &[u8], message: &[u8], ) -> Result<(), Error>

Verifies a signature on a message (with context).

This is the “Ed448” mode of RFC 8032 (no pre-hashing, a context is provided). The context string MUST have length at most 255 bytes. Return value is Ok on a valid signature, Error otherwise.

Note: this function is not constant-time; it assumes that the public key and signature value are public data.

pub fn verify_prehashed<D>( self, sig: &Signature, ctx: Option<&[u8]>, prehashed_message: D, ) -> Result<(), Error>
where D: PreHash,

Verifies a signature on a hashed message.

This is the “Ed448ph” mode of RFC 8032 (message is pre-hashed), also known as “HashEdDSA on Curve448”. The hashed message prehashed_message is provided (presumably, that hash value was obtained with SHAKE256 and a 64-byte output; the caller does the hashing itself). A context string ctx is also provided; it MUST have length at most 255 bytes. Return value is Ok on a valid signature, Error otherwise.

Note: this function is not constant-time; it assumes that the public key and signature value are public data.

fn verify_inner( &self, signature: &Signature, phflag: u8, ctx: &[u8], m: &[u8], ) -> Result<(), Error>